feat: add support for importing functionality from other libs
